Graph formatting:
The data is now displayed correctly on the graph and the user can input more than once a day. This has been changed to once a second.

Ps - need to add a stack to reduce the total number of stock values displayed on the graph - this will become an issue with prolonged program use.