Web Login page:
This is a dev record for future dates. Bcrypt is not a lite hashing algorithm. Thus, would be unsecure when carrying out cryptographic operations.

I have used ChatGPT to create a client and server. server does the cryptographic operations and the data is sent back to the client. - This will be shown in the next push