Web Login page V2:
The login page is now built on node.js. This has allowed me to use SQLite and therefore, removing the need for me to export my database into a JSON file.

The node.js file uses Express.js  to simplify the process of building web and mobile applications by providing a set of tools and features for handling routing, middleware, requests, responses, and more. It enhances productivity, flexibility, and scalability. For example: the json response - code 200 would mean the function was successful (true). This would allow the program to switch pages.

The node also uses CORS. CORS or Cross-Origin Resource Sharing in Node. js is a mechanism by which a front-end client can make requests for resources to an external back-end server

Finally, the node uses bcrypt to compare the input password (in the html file) to the password hash within the erp_system database. Carrying out this check on a node.js server ensures the security of the users sensitive information (passwords etc)